+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
 * Adds a CSV reader to process [SNOWEX](https://nsidc.org/data/snex23_ssa/versions/1) granules. (Issue-108)
+* Handle a bug in `xarray` when processing NSIDC-0630 v2 granules. (Issue-152)
 
 ## v1.3.0