+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## UNRELEASED
 
-* Adds a CSV reader to process [SNOWEX](https://nsidc.org/data/snex23_ssa/versions/1) granules. (Issue-108)
-* Refines regex handling for multi-data-file granules. (Issue-103)
+* Add a CSV reader to process [SNOWEX](https://nsidc.org/data/snex23_ssa/versions/1) granules. (Issue-108)
+* Handle a bug in `xarray` when processing NSIDC-0630 v2 granules. (Issue-152)
+* Refine regex handling for multi-data-file granules. (Issue-103)
 
 ## v1.3.0