+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Adds a CSV reader to process [SNOWEX](https://nsidc.org/data/snex23_ssa/versions/1) granules. (Issue-108)
 * Handle a bug in `xarray` when processing NSIDC-0630 v2 granules. (Issue-152)
+* Support creating release-candidates of MetGenC (v1.4.0rc1, 1.4.0rc2, ...) (Issue-128)
 
 ## v1.3.0