+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v1.4.0rc0
+## v1.4.0rc1
 
 * Adds a CSV reader to process [SNOWEX](https://nsidc.org/data/snex23_ssa/versions/1) granules. (Issue-108)
 * Handle a bug in `xarray` when processing NSIDC-0630 v2 granules. (Issue-152)